TIP-835: Allow it to work in EE
Fix dump command for EE

Add wrappers to pass cucumber

Fix paths in package.json

Cleanup

Cleanup webpack test config

Fix formatting

Fix html output in webpack test config
