+++ tests/front/acceptance/cucumber/step-definitions/structure/locales.js
@@ -1,12 +1,14 @@
-const {Given, Then} = require('cucumber');
-const {createLocale} = require('../fixtures');
-const {answerJson, csvToArray} = require('../tools');
+module.exports = function(cucumber) {
+    const {Given} = cucumber;
+    const {createLocale} = require('../../fixtures');
+    const {answerJson, csvToArray} = require('../../tools');
 
-Given('the locales {string}', async function(csvLocaleCodes) {
-  const locales = csvToArray(csvLocaleCodes).map(localeCode => createLocale(localeCode));
-  this.page.on('request', request => {
-    if (request.url().includes('/configuration/locale/rest')) {
-      answerJson(request, locales);
-    }
-  });
-});
+    Given('the locales {string}', async function(csvLocaleCodes) {
+        const locales = csvToArray(csvLocaleCodes).map(localeCode => createLocale(localeCode));
+        this.page.on('request', request => {
+            if (request.url().includes('/configuration/locale/rest')) {
+                answerJson(request, locales);
+            }
+        });
+    });
+};
