+++ webpack/test/acceptance/run.js
@@ -1,65 +1,12 @@
-var reporter = require('cucumber-html-reporter');
-const argv = require('yargs').argv
+const path = require('path');
 const cucumber = require('cucumber');
-const VError = require('verror');
+const base = process.cwd();
 
-function exitWithError(error) {
-  console.error(VError.fullStack(error)) // eslint-disable-line no-console
-  process.exit(1)
-}
+const StepDictionary = require('step-dictionary');
+const world = path.resolve(base, './tests/front/acceptance/cucumber/world.js');
+const community = path.resolve(base, './tests/front/acceptance/cucumber/step-definitions');
+const dictionary = new StepDictionary(community);
 
-const reportParameters = true === argv.report ? ['-f', 'json:coverage/acceptance-js.json'] : [];
-const featureParameters = 1 === argv._.length ? [argv._[0]] : ['./tests/front/acceptance/features'];
-const worldParameters = ['--world-parameters', JSON.stringify({
-  debug: true === argv.debug,
-  random: argv.random || true,
-  maxLatency: argv.maxLatency || 1000
-})];
+require(world)(cucumber);
+dictionary.paths.forEach(file => require(file)(cucumber));
 
-const defaultParameters = [
-  process.argv[0],
-  process.argv[1],
-  '-r',
-  './tests/front/acceptance/cucumber',
-  '-f',
-  'progress'
-];
-
-const parameters = [...defaultParameters, ...worldParameters, ...featureParameters, ...reportParameters]
-
-const run = async function () {
-  const cwd = process.cwd()
-  const cli = new cucumber.Cli({
-    argv: parameters,
-    cwd,
-    stdout: process.stdout,
-  })
-
-  let result
-  try {
-    result = await cli.run()
-  } catch (error) {
-    exitWithError(error)
-  }
-
-  const exitCode = result.success ? 0 : 1
-  if (result.shouldExitImmediately) {
-    process.exit(exitCode)
-  } else {
-    process.exitCode = exitCode
-  }
-}
-
-run();
-
-if (true === argv.report) {
-  var options = {
-      theme: 'bootstrap',
-      jsonFile: 'coverage/acceptance-js.json',
-      output: 'coverage/acceptance/cucumber_report.html',
-      reportSuiteAsScenarios: true,
-      launchReport: true
-  };
-
-  reporter.generate(options);
-}
